requirements.txt
Flask==2.3.3
google-cloud-language==2.11.1
google-auth==2.23.4
google-cloud-core==2.3.3
protobuf==4.24.4
Werkzeug==2.3.7
gunicorn==21.2.0